Change return type to getBalance function

+++ hedera-mirror-test/src/test/resources/solidity/contracts/EquivalenceContract.sol
@@ -53,7 +53,7 @@ contract EquivalenceContract {
         return abi.decode(result, (address));
     }
 
-    function getBalance(address _address) external view returns (uint) {
+    function getBalance(address _address) external view returns (uint256) {
         return _address.balance;
     }
